requirements.txt
python-dotenv==1.0.1
websockets==13.1
mysql-connector-python==9.1.0
